✅ move parse_help_output to test/helper.py
