+++ tests/cli/ems/test_cli_ems_main.py
@@ -4,7 +4,7 @@ Testing the ems cli app
 from typer.testing import CliRunner
 
 from fotoobo.cli.main import app
-from fotoobo.helpers.cli import parse_help_output
+from tests.helper import parse_help_output
 
 runner = CliRunner()
