+++ tests/cli/faz/test_cli_faz_get.py
@@ -6,8 +6,7 @@ from unittest.mock import MagicMock, patch
 from typer.testing import CliRunner
 
 from fotoobo.cli.main import app
-from fotoobo.helpers.cli import parse_help_output
-from tests.helper import ResponseMock
+from tests.helper import ResponseMock, parse_help_output
 
 runner = CliRunner()
