+++ tests/cli/fgt/test_cli_fgt_check.py
@@ -7,8 +7,7 @@ from _pytest.monkeypatch import MonkeyPatch
 from typer.testing import CliRunner
 
 from fotoobo.cli.main import app
-from fotoobo.helpers.cli import parse_help_output
-from tests.helper import ResponseMock
+from tests.helper import ResponseMock, parse_help_output
 
 runner = CliRunner()
