+++ tests/cli/fgt/test_cli_fgt_confcheck.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock, patch
 from typer.testing import CliRunner
 
 from fotoobo.cli.main import app
-from fotoobo.helpers.cli import parse_help_output
+from tests.helper import parse_help_output
 
 runner = CliRunner()
