+++ tests/cli/fmg/test_cli_fmg_get.py
@@ -8,8 +8,7 @@ from typer.testing import CliRunner
 
 from fotoobo.cli.main import app
 from fotoobo.exceptions import GeneralError, GeneralWarning
-from fotoobo.helpers.cli import parse_help_output
-from tests.helper import ResponseMock
+from tests.helper import ResponseMock, parse_help_output
 
 runner = CliRunner()
