+++ tests/cli/test_cli_main.py
@@ -10,7 +10,7 @@ from _pytest.monkeypatch import MonkeyPatch
 from typer.testing import CliRunner
 
 from fotoobo.cli.main import app
-from fotoobo.helpers.cli import parse_help_output
+from tests.helper import parse_help_output
 
 runner = CliRunner()
